IGNITE-17464 Minor fixes in NetworkMessage annotation processor
* Fix an issue when multiple interfaces define the same getter
* Move non-integration tests to the unit tests folder
