+++ modules/network/src/test/java/org/apache/ignite/internal/network/processor/InMemoryJavaFileManager.java
