+++ modules/network/src/test/java/org/apache/ignite/network/messages/InheritedMessage.java
@@ -15,10 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.network.processor;
+package org.apache.ignite.network.messages;
 
-import org.apache.ignite.network.annotations.MessageGroup;
+import org.apache.ignite.network.NetworkMessage;
+import org.apache.ignite.network.annotations.Transferable;
 
-@MessageGroup(groupType = 1, groupName = "NetworkMessageProcessorTest")
-public class ItTestMessageGroup {
+/**
+ * {@link NetworkMessage} implementation.
+ */
+@Transferable(TestMessageTypes.INHERITED_MESSAGE)
+public interface InheritedMessage extends NetworkMessage2 {
+    int intZ();
 }
