+++ modules/network/src/test/java/org/apache/ignite/network/messages/MessageWithMarshallable.java
@@ -15,14 +15,17 @@
  * limitations under the License.
  */
 
+package org.apache.ignite.network.messages;
 
-package org.apache.ignite.internal.network.processor;
-
+import java.util.Map;
+import org.apache.ignite.internal.network.serialization.SimpleSerializableObject;
 import org.apache.ignite.network.NetworkMessage;
+import org.apache.ignite.network.annotations.Marshallable;
+import org.apache.ignite.network.annotations.Transferable;
 
-/**
- * {@link NetworkMessage} implementation.
- */
-interface NetworkMessage1 extends NetworkMessage {
-    int intY();
+/** Message with a marshallable field. */
+@Transferable(TestMessageTypes.TEST_MESSAGE_WITH_MARSHALLABLE)
+public interface MessageWithMarshallable extends NetworkMessage {
+    @Marshallable
+    Map<String, SimpleSerializableObject> marshallableMap();
 }
