+++ modules/network/src/test/java/org/apache/ignite/network/messages/NetworkMessage2.java
@@ -15,16 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.network.processor;
 
+package org.apache.ignite.network.messages;
+
+import java.io.Serializable;
 import org.apache.ignite.network.NetworkMessage;
-import org.apache.ignite.network.TestMessageTypes;
-import org.apache.ignite.network.annotations.Transferable;
 
 /**
  * {@link NetworkMessage} implementation.
  */
-@Transferable(TestMessageTypes.INHERITED_MESSAGE)
-public interface InheritedMessage extends NetworkMessage2 {
-    int intZ();
+interface NetworkMessage2 extends NetworkMessage1, Serializable {
+    int intX();
 }
