+++ modules/network/src/test/resources/org/apache/ignite/internal/network/processor/AllTypesMessage.java
@@ -15,18 +15,67 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.network;
+package org.apache.ignite.internal.network.processor;
 
 import java.io.Serializable;
+import java.util.BitSet;
+import java.util.Collection;
 import java.util.Map;
+import java.util.UUID;
+import org.apache.ignite.lang.IgniteUuid;
+import org.apache.ignite.network.NetworkMessage;
 import org.apache.ignite.network.annotations.Transferable;
 
 /**
  * {@link NetworkMessage} implementation.
  */
-@Transferable(TestMessageTypes.TEST)
-public interface TestMessage extends NetworkMessage, Serializable {
-    String msg();
+@Transferable(1)
+public interface AllTypesMessage extends NetworkMessage, Serializable {
+    byte byteA();
 
-    Map<Integer, String> map();
+    short shortB();
+
+    int intC();
+
+    long longD();
+
+    float floatE();
+
+    double doubleF();
+
+    char charG();
+
+    boolean booleanH();
+
+    byte[] byteArrI();
+
+    short[] shortArrJ();
+
+    int[] intArrK();
+
+    long[] longArrL();
+
+    float[] floatArrM();
+
+    double[] doubleArrN();
+
+    char[] charArrO();
+
+    boolean[] booleanArrP();
+
+    String strQ();
+
+    BitSet bitSetR();
+
+    UUID uuidS();
+
+    IgniteUuid igniteUuidT();
+
+    NetworkMessage netMsgU();
+
+    NetworkMessage[] netMsgArrV();
+
+    Collection<NetworkMessage> netMsgCollW();
+
+    Map<String, NetworkMessage> newMsgMapX();
 }
