+++ modules/network/src/test/resources/org/apache/ignite/internal/network/processor/UnmarshallableTypeNonSerializableMessage.java
